requirements.txt
discord.py
pandas
matplotlib
requests
flask